Удалена дефолтная реализация сбокщика пакетов и заменена на реализацию через stream
